Created login/signup scene in Unity
Created a "Welcome" scene that allows users to signup and login to DIGital. Credentials are stored in a local pgAdmin database for now, but those will be moved to a remote SQL server later.

Used the exisiting auth-api .js code for the login and signup routes, and then created a new AccoutnManagement.cs file that is the middleman between Unity and the .js code.

Still need to add success messages for created accounts and login, and also create a transition to a new scene once the user is logged in.